Implemented player status, event box, and roll button.

+++ Position.pde
@@ -1,18 +1,18 @@
 class Position {
-  private int x;
-  private int y;
+  private float x;
+  private float y;
 
   /**
   * Represents a position in 2D space using X and Y
   */
-  public Position(int x, int y) {
+  public Position(float x, float y) {
     this.x=x;
     this.y=y;
   }
 
   //returns the x value
-  public int getX() {return x;}
+  public float getX() {return x;}
 
   //returns the y value
-  public int getY() {return y;}
+  public float getY() {return y;}
 }
